+++ impactlab_tools/utils/files.py
@@ -9,7 +9,7 @@ import os
 import yaml
 
 
-SHAREDDIR_SHELLVAR = 'IMPERICS_SHAREDDIR'  # Shell variable name pointing to server.yml
+SHAREDDIR_SHELLVAR = "IMPERICS_SHAREDDIR"
 default_server_config_path = "../server.yml"
 server_config = None
 
@@ -23,12 +23,13 @@ def sharedpath(subpath):
 
         default_path = os.environ.get(SHAREDDIR_SHELLVAR)
         if default_path is None:
-            # No shell (environment) variable set. Old behavior
+            # No shell (environment) variable set. Old behavior.
             default_path = str(default_server_config_path)
             msg = "Cannot find configuration file at {}".format(default_path)
             assert os.path.exists(default_path), msg
             server_config_dict = get_file_config(default_path)
         else:
+            # Use shell (environment) variable to get "shareddir".
             server_config_dict = {shareddir_key: str(default_path)}
 
         use_config(server_config_dict)